1.0.29 Execute support
Execute vanilla command now supports psudo commands. The old behavior with only psudo still works, except for psudoas. The equivalent to "psudoas <target> <command>" is "execute as <target> run psudoas <command>".

With execute, selectors (e.g. @s) and coordinates (e.g. for distance=) from any psudo depends now on the last "as" of the execute command. To make someone run the command after psudo, this should be psudoas and not psudo and this person must be selected in the last "as" of execute command.

Don't forget to use psudouuid with mobs to replace the selector by the uuid and not a poor "pig", even with @s !

(Based on the other fork https://github.com/Numa-Lab/PsudoCommands)