Add command psudoasop
The new command psudoasop is like psudoas except for plugin commands. If the following command is a plugin command, the pre-check permissions will be ignored, i.e. the permission set to the command in the plugin.yml is ignored so the player can perform the command, so the command DOES NOT give operator permission. If the plugin command checks permissions inside onCommand method, then psudoasop will still fail if the sender hasn't the permissions and I can't do anything for this, developers should know how to link permissions to a command.

Be careful with psudoasop, for now timings don't work.